Refactor: modularize helper modules into dedicated files
